--- data.csv
@@ -1,1 +1,0 @@
-user_id,name,login,email,password
